test-requirements.txt
jsonschema
boto3
pytest
mock
coverage
flake8
pytest-mock
awsiotsdk 
influxdb-client